fix: refactor sidebar
